--- 2023_balance_checks.csv
@@ -1,1 +1,0 @@
-date,account,amount,match?
