--- 2023_income.csv
@@ -1,1 +1,0 @@
-date,amount,source
