--- 2023_transaction_history.csv
@@ -1,1 +1,0 @@
-date,name,account,budget,amount
