+++ 2024_income.csv
@@ -1,0 +1,1 @@
+date,amount,source
