--- presets_old.csv
@@ -1,10 +1,0 @@
-filepath,'2023_transaction_history.csv'
-checkFilename,'2023_balance_checks.csv'
-paycheckFilename,'2023_income.csv'
-accounts,[]
-budgets,[]
-employer,''
-paycheck_split,[]
-round_budget,''
-monthly_rent,0
-monthly_internet,0
